+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## HEAD
+
+* `core` - Fetch storage object lazily in `Shrine` instance (@janko)
+
 ## 3.0.0.beta2 (2019-09-11)
 
 * `column` – Allow `Attacher#load_column` to receive a hash (@janko)